As per the discussion, this changes the log implementation to use a static inline function in C implementation for the two parameter variant, and introduces a new math.ln built-in. Execution tests have been added for all the new math functions. There is one remaining problem with this implementation in that it breaks the Lua translator test.
As discussed, the easy solution of inserting math.ln = math.log at the
top of the generated Lua files was implemented here. To avoid breaking
line counts, the command is injected at the very beginning of the file
without a newline. Unfortunately, the Lua translator test has correctly
identified that this injection was not in the original code. I'm
committing this so this can be reviewed by others so we can determine
what to do next.